moved some unnecessary structs out of data module
