+++ .travis.yml
@@ -20,3 +20,10 @@ deploy:
       branch: master
       condition: $TRAVIS_PULL_REQUEST = "false"
     script: scripts/deploy.sh
+
+  - provider: pages
+    skip-cleanup: true
+    github-token: $GITHUB_TOKEN
+    keep-history: true
+    on:
+      branch: master
