+++ README.md
@@ -1,8 +1,7 @@
 ![build status](https://travis-ci.org/awestlake87/sc2-rs.svg?branch=master)
 ![crates version](https://img.shields.io/crates/v/sc2.svg)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
-[![docs](https://docs.rs/sc2/badge.svg)](https://docs.rs/sc2)
 
-[Documentation](https://docs.rs/sc2)
+[Documentation](https://awestlake87.github.io/sc2-rs/)
 
 Rust implementation of StarCraft II Client API
